Refactored coin fetch and display process within dashboard route.
