+++ public/js/coinranking.js
@@ -15,7 +15,7 @@ async function getCoinData() {
     "https://coinranking1.p.rapidapi.com/coins?referenceCurrencyUuid=yhjMzLPhuIDl&timePeriod=24h&tiers%5B0%5D=1&orderBy=marketCap&orderDirection=desc&limit=50&offset=0",
     options
   );
-  //console.log(data.data.data.coins);
+  
   return data.data.data.coins;
 }
 
@@ -33,7 +33,7 @@ async function reducedCoinData() {
       }
     }
   }
-  //console.log(newArray);
+
   return newArray;
 }
