+++ views/homepage.handlebars
@@ -28,10 +28,11 @@
             <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
           </div>
           <div class="modal-body">
-            <form class="login-form">
+            <form class="login-form" >
               <div class="form-group">
                 <label for="username" class="modal-labels">Username</label>
                 <input type="username" class="form-control" id="username" placeholder="Username">
+                
               </div>
               <div class="form-group">
                 <label for="password" class="modal-labels">Password</label>
